requirements.txt
pandas == 2.2.3
Flask == 3.1.0
marshmallow == 4.0.0
requests == 2.32.3
flask-cors == 6.0.0
Flask-RESTful == 0.3.10
psycopg2-binary == 2.9.10
Flask-SQLAlchemy == 3.1.1